Wishlist Page Update (WIP: Loading errors gone)
